Refactor stager_app.dart to more cleanly handle single-scene apps
