+++ example/lib/post_card/post_card_scenes.stager_app.dart
@@ -8,7 +8,7 @@ import 'post_card_scenes.dart';
 
 import 'package:stager/stager.dart';
 
-Future<void> main() async {
+void main() {
   final scenes = [
     PostCardScene(),
     PostsListScene(),
