+++ lib/src/stager_app_generator.dart
@@ -44,7 +44,7 @@ $importsString
 
 import 'package:stager/stager.dart';
 
-Future<void> main() async {
+void main() {
   final scenes = [
     $sceneConstructorsString
   ];
